requirements.txt
python-telegram-bot[job-queue]
httpx
beautifulsoup4
lxml
